// notification_service/types/proto/notification.ts
// Code generated by protoc-gen-ts_proto. DO NOT EDIT.
// versions:
//   protoc-gen-ts_proto  v2.7.7
//   protoc               v3.21.12
// source: proto/notification.proto

/* eslint-disable */
import { GrpcMethod, GrpcStreamMethod } from "@nestjs/microservices";
import { Observable } from "rxjs";

export const protobufPackage = "notification";

export interface NotificationRequest {
  title: string;
  message: string;
  priority: string;
  metadata: string;
  userId: string;
  type: string;
  purpose: string;
  referenceId?: string | undefined;
  referenceType?: string | undefined;
  channels: string[];
}

export interface NotificationResponse {
  success: boolean;
  message: string;
}

export const NOTIFICATION_PACKAGE_NAME = "notification";

export interface NotificationServiceClient {
  createNotification(request: NotificationRequest): Observable<NotificationResponse>;
}

export interface NotificationServiceController {
  createNotification(
    request: NotificationRequest,
  ): Promise<NotificationResponse> | Observable<NotificationResponse> | NotificationResponse;
}

export function NotificationServiceControllerMethods() {
  return function (constructor: Function) {
    const grpcMethods: string[] = ["createNotification"];
    for (const method of grpcMethods) {
      const descriptor: any = Reflect.getOwnPropertyDescriptor(constructor.prototype, method);
      GrpcMethod("NotificationService", method)(constructor.prototype[method], method, descriptor);
    }
    const grpcStreamMethods: string[] = [];
    for (const method of grpcStreamMethods) {
      const descriptor: any = Reflect.getOwnPropertyDescriptor(constructor.prototype, method);
      GrpcStreamMethod("NotificationService", method)(constructor.prototype[method], method, descriptor);
    }
  };
}

export const NOTIFICATION_SERVICE_NAME = "NotificationService";